refactor: Refactor components to reflect 0425 PR review
Modal, Card, Header 일부 컴포넌트 및 svg 부분 새파일로 분리

ThumbNail props의 thumb_images에 직접 접근, useEffect 간소화

fetch API 사용시 catch 에러 처리 조항 추가

null return 삼항연산자 전부 ||로 변경, clickHandler click{대상}으로 재명명
