+++ src/js/components/common/Button.jsx
@@ -23,10 +23,9 @@ const ButtonWrapper = styled.button`
 	line-height: 26px;
 	color: ${(props) => (props.isAble ? "#fff" : "#BDBDBD")};
 `;
-const Button = ({ isAble, clickHandler }) => (
-	<ButtonWrapper isAble={isAble} onClick={isAble ? clickHandler : null}>
+const Button = ({ isAble, onClick }) => (
+	<ButtonWrapper isAble={isAble} onClick={!isAble || onClick}>
 		{isAble ? "주문하기" : "일시품절"}
 	</ButtonWrapper>
 );
-
 export default Button;
