+++ src/js/components/common/Card.jsx
@@ -1,5 +1,6 @@
 import { useState } from "react";
 import styled, { ThemeProvider } from "styled-components";
+import Price from "./CardPrice";
 import Badge from "./Badge";
 
 const theme = {
@@ -8,13 +9,13 @@ const theme = {
 		MEDIUM: "308px",
 	},
 	hoverTop: {
-		LARGE : "140px",
-		MEDIUM : "102px",
+		LARGE: "140px",
+		MEDIUM: "102px",
 	},
 	hoverLeft: {
-		LARGE : "144px",
-		MEDIUM : "106px",
-	}
+		LARGE: "144px",
+		MEDIUM: "106px",
+	},
 };
 
 const CardWrapper = styled.li`
@@ -27,14 +28,14 @@ const CardWrapper = styled.li`
 //스타일 컴포넌트의 네이밍은 피그마를 참고했습니다
 const CardImage = styled.img`
 	position: relative;
-	width:${({ theme: { sizes }, size }) => sizes[size]};
-	height:${({ theme: { sizes }, size }) => sizes[size]};
+	width: ${({ theme: { sizes }, size }) => sizes[size]};
+	height: ${({ theme: { sizes }, size }) => sizes[size]};
 	border-radius: 5px;
 	filter: ${(props) => (props.isHover ? "brightness(40%)" : "brightness(100%)")};
 	margin-bottom: 16px;
 `;
 const CardInfo = styled.div`
-	width:${({ theme: { sizes }, size }) => sizes[size]};
+	width: ${({ theme: { sizes }, size }) => sizes[size]};
 	display: flex;
 	flex-direction: column;
 	align-items: flex-start;
@@ -55,35 +56,11 @@ const CardBody = styled.div`
 	line-height: 20px;
 	color: #828282;
 `;
-const CardPrice = styled.div`
-	display: flex;
-	flex-direction: row;
-	align-items: flex-end;
-	padding: 0px;
-	position: static;
-	height: 29px;
-	left: 0px;
-	top: 467px;
-	margin-bottom: 16px;
-`;
-const CardPrice1 = styled.div`
-	font-weight: bold;
-	font-size: 20px;
-	line-height: 29px;
-	color: #010101;
-	margin-right: 8px;
-`;
-const CardPrice2 = styled.div`
-	font-size: 14px;
-	line-height: 20px;
-	text-decoration-line: line-through;
-	color: #bdbdbd;
-`;
 const CardHover = styled.ul`
 	position: absolute;
-	top:${({ theme: { hoverTop }, size }) => hoverTop[size]};
-	left:${({ theme: { hoverLeft }, size }) => hoverLeft[size]};
-	height:103px;
+	top: ${({ theme: { hoverTop }, size }) => hoverTop[size]};
+	left: ${({ theme: { hoverLeft }, size }) => hoverLeft[size]};
+	height: 103px;
 	display: flex;
 	align-items: flex-start;
 	justify-content: space-between;
@@ -95,38 +72,35 @@ const CardHover = styled.ul`
 	line-height: 35px;
 `;
 
-const Card = (props) => {
-	const { detail_hash, image, title, description, n_price, s_price, badge, size, delivery_type } = props;
-	const { setModalData, setModalOn } = props;
+const Card = ({ data, size, setModalOn, setModalData }) => {
+	const { detail_hash, image, title, description, n_price, s_price, badge, delivery_type } = data;
 	const [isHover, setHover] = useState(false);
 	const [src, setSrc] = useState(image);
-	const clickHandler = () => {
+	const clickImage = () => {
 		fetch(`https://h3rb9c0ugl.execute-api.ap-northeast-2.amazonaws.com/develop/baminchan/detail/${detail_hash}`)
-			.then((res) => res.json())
+			.then((response) => response.json())
 			.then((json) => {
-				if (!json.hash) return;
+				if (!json.hash) return console.log("error!! :", json);
 				json.data.name = title;
 				json.data.badge = badge;
-				setModalData(() => json);
+				setModalData(() => json.data);
 				setModalOn(true);
-			});
+			})
+			.catch((response) => console.log("error!! :", response));
 	};
 	return (
 		<ThemeProvider theme={theme}>
 			<CardWrapper>
-				<div onMouseEnter={() => setHover(true)} onMouseLeave={() => setHover(false)} onClick={clickHandler}>
+				<div onMouseEnter={() => setHover(true)} onMouseLeave={() => setHover(false)} onClick={clickImage}>
 					<CardImage src={src} size={size} isHover={isHover} onError={() => setSrc("https://codesquad.kr/img/company/codesquad2.png")}></CardImage>
-					<CardHover size={size}>{isHover ? delivery_type.map((e) => <li key={e}>{e}</li>) : ""}</CardHover>
+					<CardHover size={size}>{!isHover || delivery_type.map((e) => <li key={e}>{e}</li>)}</CardHover>
 				</div>
 				<CardInfo size={size}>
 					<CardName>{title}</CardName>
 					<CardBody>{description}</CardBody>
 				</CardInfo>
-				<CardPrice>
-					<CardPrice1>{n_price ? n_price + "원" : s_price}</CardPrice1>
-					<CardPrice2>{n_price ? s_price : ""}</CardPrice2>
-				</CardPrice>
-				{badge ? <Badge badge={badge} /> : ""}
+				<Price n_price={n_price} s_price={s_price} />
+				{!badge || <Badge badge={badge} />}
 			</CardWrapper>
 		</ThemeProvider>
 	);
